+++ src/libstddjb/socket_ioloop_send4.c
@@ -1,8 +1,10 @@
 /* ISC license. */
 
+#include <sys/types.h>
+#include <stdint.h>
 #include <skalibs/socket.h>
 
-int socket_ioloop_send4 (int fd, char *s, unsigned int len, char *ip, uint16 *port)
+ssize_t socket_ioloop_send4 (int fd, char *s, size_t len, char *ip, uint16_t *port)
 {
   return socket_send4(fd, s, len, ip, *port) ;
 }
