+++ src/libstddjb/tain_fmt.c
@@ -1,9 +1,10 @@
 /* ISC license. */
 
+#include <sys/types.h>
 #include <skalibs/fmtscan.h>
 #include <skalibs/tai.h>
 
-unsigned int tain_fmt (char *s, tain_t const *a)
+size_t tain_fmt (char *s, tain_t const *a)
 {
   char pack[TAIN_PACK] ;
   tain_pack(pack, a) ;
